--- test/valid-data/generic-mapped-complex-all/schema.json
@@ -1,26 +1,0 @@
-{
-  "$ref": "#/definitions/MyType",
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "definitions": {
-    "MyType": {
-      "additionalProperties": false,
-      "properties": {
-        "bar": {
-          "type": "string"
-        },
-        "baz": {
-          "type": "boolean"
-        },
-        "foo": {
-          "type": "string"
-        }
-      },
-      "required": [
-        "bar",
-        "baz",
-        "foo"
-      ],
-      "type": "object"
-    }
-  }
-}
