+++ test/valid-data/generic-mapped-complex/main.ts
@@ -1,0 +1,14 @@
+import { OverrideSimple } from "./util";
+
+export type Base = {
+    foo: string;
+    bar: number;
+};
+
+export type MyType = OverrideSimple<
+    Base,
+    {
+        bar: string;
+        baz: boolean;
+    }
+>;
