--- test/valid-data/generic-valueof-all/main.ts
@@ -1,8 +1,0 @@
-const RuntimeObject = {
-    FOO: "foo-val",
-    BAR: "bar-val",
-} as const;
-
-type ValueOf<T> = T[keyof T];
-
-export type MyType = ValueOf<typeof RuntimeObject>;
