--- test/valid-data/generic-valueof-all/schema-pre-fix.json
@@ -1,16 +1,0 @@
-{
-  "$ref": "#/definitions/MyType",
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "definitions": {
-    "MyType": {
-      "$ref": "#/definitions/ValueOf%3Cobject-984382748-21-65-984382748-21-74-984382748-5-74-984382748-0-74-984382748-0-75-984382748-0-160%3E"
-    },
-    "ValueOf<object-984382748-21-65-984382748-21-74-984382748-5-74-984382748-0-74-984382748-0-75-984382748-0-160>": {
-      "enum": [
-        "foo-val",
-        "bar-val"
-      ],
-      "type": "string"
-    }
-  }
-}
