--- test/valid-data/generic-valueof-star/schema-pre-fix.json
@@ -1,16 +1,0 @@
-{
-  "$ref": "#/definitions/MyType",
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "definitions": {
-    "MyType": {
-      "$ref": "#/definitions/ValueOf%3Cobject-1465228025-28-72-1465228025-28-81-1465228025-12-81-1465228025-6-81-1465228025-0-82-1465228025-0-174%3E"
-    },
-    "ValueOf<object-1465228025-28-72-1465228025-28-81-1465228025-12-81-1465228025-6-81-1465228025-0-82-1465228025-0-174>": {
-      "enum": [
-        "foo-val",
-        "bar-val"
-      ],
-      "type": "string"
-    }
-  }
-}
