--- test/valid-data/generic-valueof-star/schema.json
@@ -1,13 +1,0 @@
-{
-  "$ref": "#/definitions/MyType",
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "definitions": {
-    "MyType": {
-      "enum": [
-        "foo-val",
-        "bar-val"
-      ],
-      "type": "string"
-    }
-  }
-}
