+++ test/valid-data/generic-valueof/main.ts
