+++ test/valid-data/generic-valueof/schema.json
